Move testing code into separate executable
