+++ ffi/examples/read-table/CMakeLists.txt
@@ -1,12 +1,22 @@
 cmake_minimum_required(VERSION 3.12)
 project(read_table)
 option(PRINT_DATA "Print out the table data. Requires arrow-glib" ON)
+
+# Configuration for the read_table executable
 add_executable(read_table read_table.c arrow.c)
 target_compile_definitions(read_table PUBLIC DEFINE_DEFAULT_ENGINE)
 target_include_directories(read_table PUBLIC "${CMAKE_CURRENT_SOURCE_DIR}/../../../target/ffi-headers")
 target_link_directories(read_table PUBLIC "${CMAKE_CURRENT_SOURCE_DIR}/../../../target/debug")
 target_link_libraries(read_table PUBLIC delta_kernel_ffi)
 target_compile_options(read_table PUBLIC)
+
+# Configuration for the `test_expr executable
+add_executable(test_expr test_expr.c)
+target_compile_definitions(test_expr PUBLIC DEFINE_DEFAULT_ENGINE)
+target_include_directories(test_expr PUBLIC "${CMAKE_CURRENT_SOURCE_DIR}/../../../target/ffi-headers")
+target_link_directories(test_expr PUBLIC "${CMAKE_CURRENT_SOURCE_DIR}/../../../target/debug")
+target_link_libraries(test_expr PUBLIC delta_kernel_ffi)
+target_compile_options(test_expr PUBLIC)
 
 # Add the test
 include(CTest)
