+++ ffi/examples/read-table/expression.h
@@ -705,47 +705,3 @@ void print_tree(FILE* to, ExpressionRef ref, int depth)
       break;
   }
 }
-
-#define TEST_BUF_SIZE 4096
-
-void read_expected_expression_tree(char* expected_buf)
-{
-  FILE* data_file = fopen("expression_test_results.txt", "r");
-  if (NULL == data_file) {
-    abort();
-  }
-  if (NULL == fgets(expected_buf, TEST_BUF_SIZE - 1, data_file)) {
-    abort();
-  }
-  if (0 != fclose(data_file)) {
-    abort();
-  }
-}
-void get_expression_tree(ExpressionRef ref, char* out_buf, size_t buf_len)
-{
-
-  FILE* buf_stream = fmemopen(out_buf, buf_len, "w");
-  if (NULL == buf_stream) {
-    abort();
-  }
-  print_tree(buf_stream, ref, 0);
-  fclose(buf_stream);
-}
-void test_kernel_expr()
-{
-  SharedExpression* pred = get_kernel_expression();
-  ExpressionRef ref = construct_predicate(pred);
-
-  char out_buf[TEST_BUF_SIZE] = { 0 };
-  char expected_buf[TEST_BUF_SIZE] = { 0 };
-
-  read_expected_expression_tree(expected_buf);
-  get_expression_tree(ref, out_buf, TEST_BUF_SIZE);
-
-  for (int i = 0; i < TEST_BUF_SIZE; i++) {
-    assert(out_buf[i] == expected_buf[i]);
-  }
-
-  free_expression(ref);
-  free_kernel_predicate(pred);
-}
