+++ ffi/examples/read-table/read_table.c
@@ -3,7 +3,6 @@
 #include <string.h>
 
 #include "arrow.h"
-#include "expression.h"
 #include "read_table.h"
 #include "schema.h"
 
@@ -106,7 +105,11 @@ void scan_row_callback(
 {
   (void)size; // not using this at the moment
   struct EngineContext* context = engine_context;
-  print_diag("Called back to read file: %.*s. (size: %" PRIu64 ", num records: ", (int)path.len, path.ptr, size);
+  print_diag(
+    "Called back to read file: %.*s. (size: %" PRIu64 ", num records: ",
+    (int)path.len,
+    path.ptr,
+    size);
   if (stats) {
     print_diag("%" PRId64 ")\n", stats->num_records);
   } else {
